backend/requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.2
Automat==20.2.0
autopep8==1.5.4
cassandra-driver==3.23.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.10
django-cassandra-engine==1.6.1
django-composite-foreignkey==1.1.0
django-cors-headers==2.4.0
django-dynamic-model==0.1.0
django-excel==0.0.10
django-extensions==3.0.9
django-filter==2.2.0
django-memcached==0.1.2
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-simplejwt
djongo==1.2.38
djoser==2.1.0
docutils==0.16
ecdsa==0.15
et-xmlfile==1.0.1
geomet==0.1.2
graphene==2.1.8
graphene-django==2.10.0
graphene-mongo==0.2.12
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
iso8601==0.1.12
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.1.0
lml==0.0.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mongoengine==0.20.0
msgpack==0.6.2
numpy==1.18.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
plotly==4.8.1
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.19
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-ods==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
PyHamcrest==2.0.1
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.1
python-jose==3.1.0
python-keycloak-client==0.2.3
python-memcached==1.59
python3-openid==3.2.0
pytz==2019.3
pyucum==0.1.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
scipy
seaborn==0.10.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
statistics==1.0.3.5
texttable==1.6.2
toml==0.10.1
tqdm==4.43.0
#Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface
autopep8==1.5.4
Cython==0.29.21
et-xmlfile==1.0.1
jdcal==1.4.1
owlready2
pycodestyle==2.6.0
editdistance
orjson
django-redis
fhir.resources
fhir
